+++ fotoobo/cli/fc/__init__.py
@@ -1,7 +1,0 @@
-"""
-__init__.py
-"""
-
-from . import fc
-
-__all__ = ["fc"]