+++ fotoobo/cli/fcasset/__init__.py
@@ -1,0 +1,7 @@
+"""
+__init__.py
+"""
+
+from . import fcasset
+
+__all__ = ["fcasset"]