+++ fotoobo/inventory/inventory.py
@@ -10,7 +10,7 @@ from typing import Any, Dict, Optional
 from fotoobo.exceptions import GeneralWarning
 from fotoobo.fortinet.fortianalyzer import FortiAnalyzer
 from fotoobo.fortinet.forticlientems import FortiClientEMS
-from fotoobo.fortinet.forticloud import FortiCloudAsset
+from fotoobo.fortinet.forticloudasset import FortiCloudAsset
 from fotoobo.fortinet.fortigate import FortiGate
 from fotoobo.fortinet.fortimanager import FortiManager
 from fotoobo.helpers.config import config
@@ -170,7 +170,7 @@ class Inventory:
             elif asset.get("type", "") == "forticlientems":
                 self.assets[name] = FortiClientEMS(**asset)
 
-            elif asset.get("type", "") == "forticloud":
+            elif asset.get("type", "") == "forticloudasset":
                 self.assets[name] = FortiCloudAsset(**asset)
 
             elif asset.get("type", "") == "fortianalyzer":