+++ fotoobo/tools/fcasset/__init__.py
@@ -1,5 +1,5 @@
 """
-fc tools
+fcasset tools
 """
 
 from . import get