+++ tests/cli/fcasset/test_cli_fcasset_get.py
@@ -1,5 +1,5 @@
 """
-Testing the fc get cli app
+Testing the fcasset get cli app
 """
 
 from unittest.mock import MagicMock
@@ -13,9 +13,9 @@ from tests.helper import parse_help_output
 runner = CliRunner()
 
 
-def test_cli_app_fc_get_help() -> None:
-    """Test cli help for fc get"""
-    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fc", "get", "-h"])
+def test_cli_app_fcasset_get_help() -> None:
+    """Test cli help for fcasset get"""
+    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fcasset", "get", "-h"])
     assert result.exit_code == 0
     arguments, options, commands = parse_help_output(result.stdout)
     assert not arguments
@@ -23,9 +23,9 @@ def test_cli_app_fc_get_help() -> None:
     assert set(commands) == {"products", "version"}
 
 
-def test_cli_app_fc_get_version_help() -> None:
-    """Test cli help for fc get version"""
-    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fc", "get", "version", "-h"])
+def test_cli_app_fcasset_get_version_help() -> None:
+    """Test cli help for fcasset get version"""
+    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fcasset", "get", "version", "-h"])
     assert result.exit_code == 0
     arguments, options, commands = parse_help_output(result.stdout)
     assert not arguments
@@ -33,20 +33,21 @@ def test_cli_app_fc_get_version_help() -> None:
     assert not commands
 
 
-def test_cli_app_fc_get_version(monkeypatch: MonkeyPatch) -> None:
-    """Test cli options and commands for fc get version"""
+def test_cli_app_fcasset_get_version(monkeypatch: MonkeyPatch) -> None:
+    """Test cli options and commands for fcasset get version"""
     monkeypatch.setattr(
-        "fotoobo.fortinet.forticloud.FortiCloudAsset.post",
+        "fotoobo.fortinet.forticloudasset.FortiCloudAsset.post",
         MagicMock(return_value={"version": "3.0"}),
     )
-    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fc", "get", "version"])
+    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fcasset", "get", "version"])
     assert result.exit_code == 0
-    assert "forticloud │ 3.0" in result.stdout
+    assert "forticloudasset " in result.stdout
+    assert " │ 3.0" in result.stdout
 
 
-def test_cli_app_fc_get_products_help() -> None:
+def test_cli_app_fcasset_get_products_help() -> None:
     """Test cli help for fc get products"""
-    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fc", "get", "products", "-h"])
+    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fcasset", "get", "products", "-h"])
     assert result.exit_code == 0
     arguments, options, commands = parse_help_output(result.stdout)
     assert not arguments