+++ tests/cli/fcasset/test_cli_fcasset_main.py
@@ -1,5 +1,5 @@
 """
-Testing the fc cli app
+Testing the fcasset cli app
 """
 
 from typer.testing import CliRunner
@@ -10,9 +10,9 @@ from tests.helper import parse_help_output
 runner = CliRunner()
 
 
-def test_cli_app_fc_help() -> None:
-    """Test cli help for fc"""
-    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fc", "-h"])
+def test_cli_app_fcasset_help() -> None:
+    """Test cli help for fcasset"""
+    result = runner.invoke(app, ["-c", "tests/fotoobo.yaml", "fcasset", "-h"])
     assert result.exit_code == 0
     arguments, options, commands = parse_help_output(result.stdout)
     assert not arguments