+++ tests/tools/fcasset/__init__.py
@@ -1,0 +1,3 @@
+"""
+Here we test the fcasset tools module
+"""