+++ tests/tools/fcasset/get/test_version.py
@@ -1,4 +1,4 @@
-"""Test fc tools get version"""
+"""Test fcasset tools get version"""
 
 from pathlib import Path
 from unittest.mock import MagicMock
@@ -6,7 +6,7 @@ from unittest.mock import MagicMock
 import pytest
 from _pytest.monkeypatch import MonkeyPatch
 
-from fotoobo.tools.fc.get import version
+from fotoobo.tools.fcasset.get import version
 
 
 @pytest.fixture(autouse=True)
@@ -19,9 +19,9 @@ def inventory_file(monkeypatch: MonkeyPatch) -> None:
 
 @pytest.fixture(autouse=True)
 def fc_login(monkeypatch: MonkeyPatch) -> None:
-    """Mock the FortiCloud login to always return 200 without to really login"""
+    """Mock the FortiCloudAsset login to always return 200 without to really login"""
     monkeypatch.setattr(
-        "fotoobo.fortinet.forticloud.FortiCloudAsset.login",
+        "fotoobo.fortinet.forticloudasset.FortiCloudAsset.login",
         MagicMock(return_value=200),
     )
 
@@ -29,9 +29,9 @@ def fc_login(monkeypatch: MonkeyPatch) -> None:
 def test_version(monkeypatch: MonkeyPatch) -> None:
     """Test get version"""
     monkeypatch.setattr(
-        "fotoobo.fortinet.forticloud.FortiCloudAsset.get_version",
+        "fotoobo.fortinet.forticloudasset.FortiCloudAsset.get_version",
         MagicMock(return_value="3.0"),
     )
-    result = version("forticloud")
-    data = result.get_result("forticloud")
+    result = version("forticloudasset")
+    data = result.get_result("forticloudasset")
     assert data == "3.0"