+++ WHATSNEW.md
@@ -1,10 +1,11 @@
 ### Added
 
-- Add FortiCloud module for cloud asset management
+- Add FortiCloud module for FortiCloud Asset Management
 
 ### Changed
 
-- ems get workgroups now displays group id in table
+- CLI command "ems get workgroups" now displays group id in table
+- Tool "ems get workgroups" has changed its return format
 
 ### Fixed